Inline team crest component
